[PATCH v2] shlwapi: Fix the return value of SHAddDataBlock
Patch resent because of "formatting" issue.

This patch fixes the return type and value of SHAddDataBlock.
The corresponding test (shlwapi's clist) is adjusted to reflect
those changes. Tested on Windows 2003 and 7.

In addition I set, in SHRemoveDataBlock, the lpList pointer
to NULL, to respect the fact it's a pointer, not just a number.

Signed-off-by: Hermes Belusca-Maito <hermes.belusca@sfr.fr>